Add solution for problem `E2 - Stars Drawing (Hard Edition)` Link: https://codeforces.com/contest/1015/problem/E2 Tags: binary search, dp, greedy, *1900 Ref: https://codeforces.com/contest/1015/submission/139716568

+++ codeforces/1015/E2.cpp
@@ -1,7 +1,7 @@
 // https://github.com/AkshatAggarwal14/Competetive-Programming
 #include "bits/stdc++.h"
 using namespace std;
-using ll = int;
+using ll = int64_t;
 auto sz = [](const auto &container) -> ll { return container.size(); };
 
 int main() {
